% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{appendEachElement}
\alias{appendEachElement}
\title{Append elements of an alphabet to construct words}
\usage{
appendEachElement(L, a)
}
\arguments{
\item{L}{character, an alphabet, e.g. c("A", "C", "G", "T")}

\item{a}{numeric, desired length of the words}
}
\description{
Append elements of an alphabet to construct words
}
