% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateLists}
\alias{generateLists}
\title{Tests if a set of words contains a query at an specified hamming distance}
\usage{
generateLists(
  alphabet,
  word_length,
  required_words,
  min_dist,
  seed_word = NULL
)
}
\arguments{
\item{alphabet}{character, an alphabet, e.g. c("A", "C", "G", "T")}

\item{word_length}{numeric, desired length of the words}

\item{required_words}{numeric, number of required words}

\item{min_dist}{numeric, minimum distance to be satisfied}

\item{seed_word}{character, optional parameter to use as the first word (it can cause an infinite loop)}
}
\description{
Tests if a set of words contains a query at an specified hamming distance
}
