% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testElement}
\alias{testElement}
\title{Tests if a set of words contains a query at an specified hamming distance}
\usage{
testElement(min_dist, query, word_list)
}
\arguments{
\item{min_dist}{numeric, minimum distance to be satisfied}

\item{query}{character, a word}

\item{word_list}{list, list of words}
}
\description{
Tests if a set of words contains a query at an specified hamming distance
}
