% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{words}
\alias{words}
\title{Generate a list of all possible combinations of letters of an specified length}
\usage{
words(alph, len)
}
\arguments{
\item{alph}{character, an alphabet, e.g. c("A", "C", "G", "T")}

\item{len}{numeric, desired length of the words}
}
\description{
Generate a list of all possible combinations of letters of an specified length
}
